requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
PyTango>=9.3.3
python-dotenv>=0.19.0
pydantic>=1.8.2 